requirements.txt
# ─────────────────────────────────────────────
# Medical RAG System — Dependencies
# Python 3.10+
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────

# Data Processing
pandas
requests

# Search & Retrieval
rank_bm25
sentence-transformers
torch

# LLM & RAG
groq
python-dotenv

# XML processing is part of the standard library (xml.etree.ElementTree), 
# so it doesn't need to be listed here.